+++ doc/CHANGELOG.md
@@ -26,6 +26,8 @@ See doc/COPYRIGHT.md for more details.
 
 # Changelog
 
+* `#8419` [PDF Export] Internal error when entering wrong YAML
+
 ## 3.0.8
 
 * * `#5357` Update readme and gemspec